combined AboutTechnolgapSection.module.scss with ContactTechnolgapSection.module.scss, also renamed folder that contains the new scss file with the AboutSection.js and ContactSection.js
