--- components/AboutTechnolgapSection/AboutTechnolgapSection.module.scss
@@ -1,63 +1,0 @@
-@import "../../theme/theme";
-
-.AboutTechnolgapSection {
-    display: flex;
-    justify-content: space-around;
-    align-items: center;
-  
-    width: 100%;
-  
-    background-color: $olive-drab;
-  
-    font-family: $secondary-header-font;
-    font-size: $body-font-size;
-    color: $white;
-}
-
-.jumbotron {
-    padding: 50px 200px 70px 200px;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-}
-
-@media only screen and (max-width: 1000px) {
-    .jumbotron {
-        padding: 50px 50px 70px 50px;
-    }
-}
-
-@media only screen and (max-width: 479px) {
-    .jumbotron {
-        padding: 30px;
-    }
-}
-
-.WITTitle {
-    font-size: 40px;
-}
-
-.WITText {
-    font-size: 25px;
-    padding: 10px 0px 0px 0px;
-}
-
-.WITButton {
-    background-color: $olive-drab;
-    font-family: $secondary-header-font;
-    border: 2px white solid;
-    padding: 25px 80px;
-    color: white;
-    font-size: 18px;
-    font-weight: bold;
-    margin: 50px 10px 10px 10px;
-}
-
-.WITButton:hover {
-    background-color: $white;
-    color: $olive-drab;
-}
-
-.centering {
-    text-align: center;
-}
